+++ pg-core/src/java/org/pg/Config.java
@@ -6,7 +6,6 @@ import org.pg.enums.ConnType;
 import org.pg.enums.SSLValidation;
 import org.pg.error.PGError;
 import org.pg.json.JSON;
-import org.pg.processor.IProcessor;
 
 import javax.net.ssl.SSLContext;
 import java.util.Map;
@@ -47,8 +46,7 @@ public record Config(
         int poolBorrowConnTimeoutMs,
         boolean useUnixSocket,
         String unixSocketPath,
-        Executor executor,
-        Map <Object, IProcessor> typeMap
+        Executor executor
 ) {
 
     public ConnType getConnType() {
@@ -101,7 +99,6 @@ public record Config(
         private boolean useUnixSocket = false;
         private String unixSocketPath = null;
         private Executor executor = Const.executor;
-        private Map<Object, IProcessor> typeMap;
 
         public Builder(final String user, final String database) {
             this.user = Objects.requireNonNull(user, "User cannot be null");
@@ -298,12 +295,6 @@ public record Config(
         }
 
         @SuppressWarnings("unused")
-        public Builder typeMap(final Map<Object, IProcessor> typeMap) {
-            this.typeMap = typeMap;
-            return this;
-        }
-
-        @SuppressWarnings("unused")
         private void _validate() {
             if (!(poolMinSize <= poolMaxSize)) {
                 throw new PGError("pool min size (%s) must be <= pool max size (%s)",
@@ -346,8 +337,7 @@ public record Config(
                     this.poolBorrowConnTimeoutMs,
                     this.useUnixSocket,
                     this.unixSocketPath,
-                    this.executor,
-                    this.typeMap
+                    this.executor
             );
         }
     }