+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - read Postgres types on startup
 - fix redundant .oid in config
+- type-map support removed
 - ?
 
 ## 0.1.36